+++ scripts/component-generator/__testfixtures__/testkit-definitions.input.js
@@ -3,6 +3,6 @@
  */
 module.exports = {
   Component: {
-    unidriver: true,
+    unidriver: true
   },
 };