+++ scripts/component-generator/__testfixtures__/testkit-definitions.output.js
@@ -3,10 +3,10 @@
  */
 module.exports = {
   Component: {
-    unidriver: true,
+    unidriver: true
   },
 
   MyNewComponent: {
-    unidriver: true,
-  },
+    unidriver: true
+  }
 };