+++ src/index.js
@@ -112,3 +112,4 @@ export { default as SegmentedToggle } from './SegmentedToggle';
 export { default as FloatingNotification } from './FloatingNotification';
 export { default as RichTextInputArea } from './RichTextInputArea';
 export { default as NoBorderInput } from './NoBorderInput';
+export { default as ColorInput } from './ColorInput';