+++ stories/ColorInput/storySettings.js
@@ -1,0 +1,13 @@
+import { Category } from '../storiesHierarchy';
+
+export const storySettings = {
+  kind: Category.SELECTION,
+  category: Category.SELECTION,
+
+  storyName: '4.8 ColorInput',
+  dataHook: 'story-color-input',
+};
+
+export const testStories = {
+  STATES: 'ColorInput states',
+};