+++ testkit/component-definitions.js
@@ -236,4 +236,10 @@ export default {
       children: <GoogleAddressInput Client={GmapsTestClient} />,
     },
   },
+
+  ColorInput: {
+    props: {
+      value: '#FFFFFF',
+    },
+  },
 };