+++ testkit/testkit-definitions.js
@@ -193,4 +193,7 @@ module.exports = {
   RichTextInputArea: { unidriver: true },
 
   NoBorderInput: {},
+  ColorInput: {
+    unidriver: true,
+  },
 };